Develop

+++ app/Http/Controllers/CarouselController.php
@@ -5,8 +5,11 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\Facades\File;
+use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\Str;
 use App\Carousel;
 use Auth;
+use App;
 
 class CarouselController extends Controller
 {
@@ -28,10 +31,16 @@ class CarouselController extends Controller
       $carousel = new Carousel;
       $imageName = time().'.'.request()->img_path->getClientOriginalExtension();
       $carousel->user_id = Auth::user()->id;
-      $carousel->img_path = 'visitor/images/carousel/'.$imageName;
+      if (App::environment('local')){
+        $carousel->img_path = 'visitor/images/carousel/'.$imageName;
+        request()->img_path->move(public_path('visitor/images/carousel'), $imageName);
+      }
+      if (App::environment('production')){
+        $url = Storage::disk('s3')->putFile('carousel', request()->img_path, 'public');
+        $carousel->img_path = Storage::disk('s3')->url($url);
+      }
       $carousel->show = request()->show;
       $carousel->save();
-      request()->img_path->move(public_path('visitor/images/carousel'), $imageName);
       return redirect()->route('admin_carousel.index')
                         ->with('success','Se agregó la foto  al carousel exitosamente');
   }
@@ -50,8 +59,17 @@ class CarouselController extends Controller
   }
 
   public function admin_delete(Carousel $carousel){
-    if(File::exists($carousel->img_path)) {
-        File::delete($carousel->img_path);
+    if (App::environment('local')){
+      if(File::exists($carousel->img_path)) {
+          File::delete($carousel->img_path);
+      }
+    }
+    if (App::environment('production')){
+      $aws_url = env('AWS_URL', false);
+      $path = Str::after($carousel->img_path, $aws_url);
+      if(Storage::disk('s3')->exists($path)){
+        Storage::disk('s3')->delete($path);
+      }
     }
     $carousel->delete();
     return redirect()->route('admin_carousel.index')
